Initial work on SQLiteDatastore
This commits implements a new SQLiteDatastore API, which implements a
data storage system on top of an SQLite database. It presents a
Preferences implementation, as well as data binding support and methods
for easily accessing the data directly. It has a reasonable level of testing.

I intend to implement subsequent commits for the remaining data bindings
types, but since this is working properly for primitive types, it seemed
like a good time to cut a commit.