SQLiteDatastore: add metadata support
There is now a special metadata table which is used to store the
SQLiteDatastore magic string and format version, and which can also be
used to store a user-provided app name and version integer. Because this
is implemented as a generalized KVP store, more metadata can be added in
the future easily.