WIP - present a method to use native file selector
Even if the Fyne file dialogs are OK, it nice to be able to use the
native file selector as it proposes some native actions as "create a
directory", use bookmarks, etc.

Zenity and KDialogs are commonly installed in Linux. This work in
progress proposes a way to use native dialogs.

NOTE: this could be extended to info dialogs, alerts, and many others.