Addressed review feedback
Fixed arrowhead rotation. It wasn't a problem with fyne 2.3.5, but would have been a problem with fyne 4.0. Now it won't be a problem.
Made DrawingArea public to support application-level testing scenarios that require simulated Drag, Mouse, and Tap events to be sent to the diagram.
Made (but commented out) a change to AnchoredText for fyne 2.4.0 so that anchored text does not scroll. This needs to be uncommented for fyne 2.4.0