fix(completionentry): calculate max size and resize correctly
The new max size calculation ensures the popup position is always below
the entry widget.

The addition of the Resize method ensures the popup is updated with the
rest of the widget upon resize.

Since now the popup doesn't go above the entry widget not all options
are rendered and the test had to be updated to select the first item as
the other ones are not visible