Change handling of clipboard paste
Process new entry text after paste operation. This prevents invalid characters (minus not as first char, chars that are not minus, radix separator, thousands separator, or digits).
This more thoroughly handles pasted chars than preious processing.
Delete isNumber as it is no longer needed and is not as thorough as processing done in SetText.